Hacky workaround to make rust-analyzer not complain about procmacro
Cargo has no idea what a "file path" during compilation.
Since we autogenerate mod statements for plugins/modules this
requires that cargo is aware of the relative path to the "modules"
directory. This causes a CRY because modules is in a sub-crate
and rust-analyzer builds from the context of the sub-crate not the
parent crate

Sidestep all this weirdness for now using more glob magic